Bug 1324581 - check event handled by apz in nsSliderFrame::StartAPZDrag r=botond
test_scroll_snapping_scrollbars.html is failing when apz.drag.enabled
is true because synthesizeMouse does not route events through APZ,
meaning that the drag is never handled by APZ or Gecko.

Resolve this by adding a check that mHandledByApz flag is true for
WidgetGUIEvent, and handle in Gecko if that is not the case.

MozReview-Commit-ID: IrxVLxqs64V

UltraBlame original commit: b51431dfd912e338fa6e520061ce64b1f0af5aba