Bug 1325148 - Use "error" level in mozleak when logging leak failures, r=jgraham
This fixes a regression when switching to the StructuredOutputParser in mozharness.
Previously, mozleak was relying on the string "TEST-UNEXPECTED-FAIL" to turn the jobs
orange, but it was doing so at the "warning" level. The StructuredOutputParser requires
the "error" level to set the appropriate tbpl_status.

MozReview-Commit-ID: 9u9mwqrkA6E

UltraBlame original commit: ee7a215666021e8736f77ab13354c93b0023fea5