Bug 1322772 - Poll for ready state on web progress stop state for image documents; r=maja_zf,whimboo
Image documents do not fire DOMContentLoaded events, but we can use the
web progress listener to call the ready state checks when the document's
progress state reaches `nsIWebProgressListener.STATE_STOP`.

This change makes it possible to navigate to image documents, such as
.jpg and .gif images with Marionette.

Documents with the text/html MIME are not affected by this because they
are parsed as HTML documents with a special style.

MozReview-Commit-ID: I92FDXDIdz9

UltraBlame original commit: 6794051829d7c818e936b1342a0fb6018720dad5