Bug 1326388 part 1.  Switch nsSVGElement from storing a StyleRule to storing a DeclarationBlock to represent its mapped attributes.  r=dbaron
The current setup complicates things by causing the existence of StyleRule
instances that have a null GetDOMRule(), which
inDOMUtils::GetCSSStyleRulesrelies on to not return the relevant rules.  Since
we want to get rid of GetDOMRule(), better to not have a StyleRule there in the
first place.

UltraBlame original commit: 13b297c37408524ae3891d550bfd28e03bfa3c76