bug 1328899 don't use GTK prelight and active state flags when drawing textfield-multiline background r=stransky+263117
This is consistent with GTK not setting these flags with GtkTextView
and GtkScrolledWindow.

The active flag was triggering selected_bg_color with Ambiance.

This patch almost reverts the state flags to the logic prior to
cd7544affe40 but keeps the focused flag in the scrolled window.

MozReview-Commit-ID: IVY1sI2Hllw

UltraBlame original commit: af6601db2875a82cbba7983858cea78455b24b74