Bug 1330822 - Remove CPOWs from browser_popupUI.js. r=Felipe
Finding the popup window after the fact was going to be messy (maintaining
window identity across the message manager is a tricky task). Instead, it's
much easier to wait for the new window to open and grab a reference to it
then.

UltraBlame original commit: a66860c81eaf8f83e38ab184c1d1fa9e5564edcc