Bug 1332573 - Always set status in mochitest even if no tests were found, r=pyang
This fixes a regression in bug 1332573 where if no tests are found by the getActiveTests
function, the status variable never gets set. This ensures that we always set status, and
that it will be set to a non-zero return code if any of the calls to runApp fail.

MozReview-Commit-ID: 20M7FcBs0DF

UltraBlame original commit: c61b0e70d0cfadd25f31fe288dbe2fb18cc7da0f