bug 1333797 - Require 'record_in_processes' values in Scalars.yaml r=Dexter
Having "main" be the default value is a very sensible choice, but one that is a
little opaque if you aren't into reading Telemetry code.

Just require a value here, and this will encourage correct use as it can be
caught and policed in data-review.

MozReview-Commit-ID: ASaWGuPAc9a

UltraBlame original commit: 43c2c4c0324de97236ccef58707fc75772a56811