Bug 1321051 - Override Supports(TrackInfo) in RemoteDecoderModule - r=jya
With e10s enabled, RemoteDecoderModule is used as a proxy, but it was not
overriding Supports(TrackInfo), meaning that some of the extra information
carried in TrackInfo (e.g.: resolution) was lost when falling back to
SupportsMimeType(nsACString). So further along, WMFDecoderModule was not able
to reject too-high resolutions.

Overriding Supports(TrackInfo) to just forward the TrackInfo data is enough to
re-enable resolution checks on Windows.

MozReview-Commit-ID: 3OsPmILItH7

UltraBlame original commit: 9635a6d92b1c9dcd4c8695dae6370982c870e935