servo: Merge #5252 - Put system memory measurements in a memory reporter (from nnethercote:system-reporter); r=jdm
Currently the system memory measurements ("resident", "vsize", etc.) are
not reported through the generic memory reporting mechanism, simply
because they pre-date that mechanism. This changeset removes that
special-casing.

One consequence of this is that previously if a platform didn't
implement one of the basic measurements, a '???' entry would be printed.
Now nothing will be printed. This is no great loss and matches what
Firefox does.

Another consequence is that the order in which the measurements are
printed is changed. I plan to fix this soon so that reports are sorted
in a more sensible fashion.

Source-Repo: https://github.com/servo/servo
Source-Revision: 29b158e9853f5c135bf853101ee74c5618491655

UltraBlame original commit: 7aeea49cd790abb794652e68d925b9066ab1b965