Bug 1336155 - Update rust repack script for 1.15.0 stable. r=mshal
Revert the bump to beta for the android builders now that 1.15.0
is in stable release.

Generalize the cargo tarball name mangling to use a regular
expression to hand the tiny revision added for the stable
release, as suggested by mshal in bug 1332759.

MozReview-Commit-ID: 2xZWpvKxY77

UltraBlame original commit: 668b300615182a466a4d4568962b19959ae069a4