Bug 1337828 - Pass unexpected test paths as is from mach_test_package_bootstrap to the test harness, r=maja_zf
This was a bug in the mach test package bootstrap that could happen in one-click loaners. If a user neglected to
pass in --flavor, this meant that instead of a nice error message out of mochitest, they'd get a cryptic
traceback.

With this fix, it will be obvious that the problem is "no tests were found".

MozReview-Commit-ID: 5LEQFVDoJrT

UltraBlame original commit: e725ba1475be5300919d3be04b7dedafa3e02f43