Bug 1312754 - Add a service to throttle certain HTTP channels. r=mayhemer
This patch adds (and hooks up) a new service so that HTTP channels marked as
"Throttleable" will periodically be Suspend()ed and Resume()d when more
important operations are going (such as a page load).

While this patch is not responsible for marking channels as "Throttleable",
the general idea is that these would be less-important channels - background
downloads, beacons, etc, and perhaps even resources known to be trackers.

MozReview-Commit-ID: HEZsxS04rRK

UltraBlame original commit: 145410f30206db4b6254b812c1a2a52fb90748e5