Bug 1335651 - Move index_paths from DockerImageTask to the base Task class. r=dustin
This does slightly change the behavior when artifacts expire, in that
if for some reason the artifact for the task that was found expired,
we don't try to get the artifact from a lower level task. In practice,
that shouldn't be a concern.

UltraBlame original commit: a0da8c39431d93d9efc043b44cf87087c16f4f53